+++ packages/atlas/src/components/_loaders/SkeletonLoaderVideoGrid/SkeletonLoaderVideoGrid.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 
 import { Grid } from '@/components/Grid'
 import { VideoTileViewer } from '@/components/_video/VideoTileViewer'
-import { repeat } from '@/utils/data'
+import { arrayFrom } from '@/utils/data'
 
 type SkeletonLoaderVideoGridProps = {
   videosCount?: number
@@ -11,7 +11,7 @@ type SkeletonLoaderVideoGridProps = {
 export const SkeletonLoaderVideoGrid: FC<SkeletonLoaderVideoGridProps> = ({ videosCount = 10, onResize }) => {
   return (
     <Grid onResize={onResize}>
-      {repeat(videosCount).map((_, idx) => (
+      {arrayFrom(videosCount).map((_, idx) => (
         <VideoTileViewer key={idx} />
       ))}
     </Grid>
