+++ packages/atlas/src/components/_video/VideoPlayer/CustomTimeline.tsx
@@ -2,7 +2,7 @@ import { clamp, round } from 'lodash-es'
 import { Dispatch, FC, MouseEvent, SetStateAction, TouchEvent, useEffect, useRef, useState } from 'react'
 import { VideoJsPlayer } from 'video.js'
 
-import { repeat } from '@/utils/data'
+import { arrayFrom } from '@/utils/data'
 import { formatDurationShort } from '@/utils/time'
 
 import {
@@ -105,7 +105,7 @@ export const CustomTimeline: FC<CustomTimelineProps> = ({
       }
 
       // get all buffered time ranges
-      const bufferedTimeRanges = repeat(buffered.length).map((_, idx) => ({
+      const bufferedTimeRanges = arrayFrom(buffered.length).map((_, idx) => ({
         bufferedStart: buffered.start(idx),
         bufferedEnd: buffered.end(idx),
       }))
