+++ packages/atlas/src/views/viewer/NftsView/NftsView.tsx
@@ -1,5 +1,6 @@
 import { FC, useState } from 'react'
 
+import { useNfts } from '@/api/hooks/nfts'
 import { OwnedNftOrderByInput } from '@/api/queries/__generated__/baseTypes.generated'
 import { SvgActionFilters } from '@/assets/icons'
 import { EmptyFallback } from '@/components/EmptyFallback'
@@ -12,7 +13,6 @@ import { NftTileViewer } from '@/components/_nft/NftTileViewer'
 import { VideoContentTemplate } from '@/components/_templates/VideoContentTemplate'
 import { useHeadTags } from '@/hooks/useHeadTags'
 import { useMediaMatch } from '@/hooks/useMediaMatch'
-import { useNfts } from '@/hooks/useNfts'
 import { useVideoGridRows } from '@/hooks/useVideoGridRows'
 import { createPlaceholderData } from '@/utils/data'
 import { SentryLogger } from '@/utils/logs'
